requirements.txt
Flask
Flask-Session
pytz
requests
sqlite3
pytest
wkt_scraper
